Pool Award Balance Fix: 1. change from "awardBalance" to "captureAwardBalance" in AppSlice 2. move graph calls from PoolInfo to 33Together so that they can be passed as Props
