+++ src/hooks/web3Context.tsx
@@ -28,12 +28,13 @@ const ALCHEMY_ID_LIST = [
 ];
 
 // this is the ethers common api key, it is rate limited somewhat
-const defaultApiKey = "https://eth-mainnet.alchemyapi.io/v2/_gg7wSSi0KMBsdKnGVfHDueq6xMB9EkC";
+const defaultApiKey = "_gg7wSSi0KMBsdKnGVfHDueq6xMB9EkC";
 
 const TEMP_ALCHEMY_IDS = [
   // "rZD4Q_qiIlewksdYFDfM3Y0mzZy-8Naf", // appleseed-temp1
   // "9GOp6SIgE0en92i3r0JSvxccZ0N2idmO", // appleseed-temp2
-  "j0QUyceqxu31tQrAQSotL2YMqmuzoGPh", // appleseed-temp3
+  // "j0QUyceqxu31tQrAQSotL2YMqmuzoGPh", // appleseed-temp3
+  defaultApiKey,
 ];
 function getAlchemyAPI(chainID: Number) {
   const randomIndex = Math.floor(Math.random() * ALCHEMY_ID_LIST.length);
